REL-4333 Apply patch from Oracle to build with system sqlite
Patch was taken unchanged from Oracle's perl-DBD;SQLite RPM.
By default, this module use the version of sqlite for build, that
it has bundled within its repo. RPM vendors have always built
it with the system-provided sqlite RPM instead, with the help of
this patch. So we should bring it back to continue building with
system sqlite.